Improve exports for the tooltip.dart
