+++ lib/src/widget/tooltip/data/tooltip_params.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:xayn_design/xayn_design.dart';
+
+typedef MessageBuilder = TooltipMessage Function(BuildContext context);
+
+class TooltipParams {
+  final String label;
+  final MessageBuilder builder;
+
+  const TooltipParams({
+    required this.label,
+    required this.builder,
+  });
+}
