+++ lib/xayn_design.dart
@@ -22,14 +22,5 @@ export 'src/widget/dots_indicator/dots_indicator.dart';
 export 'src/widget/nav_bar/nav_bar.dart';
 export 'src/widget/settings/settings.dart';
 export 'src/widget/switch/app_switch.dart';
-export 'src/widget/tooltip/application_tooltip_provider.dart';
-export 'src/widget/tooltip/messages/textual_notification.dart';
-export 'src/widget/tooltip/messages/tooltip_message.dart';
-export 'src/widget/tooltip/messages/tooltip_message_key.dart';
 export 'src/widget/tooltip/tooltip.dart';
-export 'src/widget/tooltip/tooltip_context_provider_state_mixin.dart';
-export 'src/widget/tooltip/tooltip_controller.dart';
-export 'src/widget/tooltip/tooltip_options.dart';
-export 'src/widget/tooltip/tooltip_state_mixin.dart';
-export 'src/widget/tooltip/tooltip_stateless_mixin.dart';
 export 'src/widget/unter_den_linden/unter_den_linden.dart';
